Make the view optional
The view_queue is not materialized, and thus functions as syntactic
sugar for a complex query. Furthermore, our infrastructure works poorly
with views. Therefore, this usage adds a significant maintainership
burden for little gain.

This commit removes the view from NanoMDM code paths, but it still
exists in the schema. This is purely for convenience and interactive
users.

Furthermore, I added an index used by the queue query, which should have
existed for the view. This should improve performance.

Testing:
* Tested migration - see the schema migration script
* Tested base functionality
* Passes new integration test