Add GitHub action for MySQL backend
Currently the MySQL backend is not tested, as it requires MySQL server
to run. I have added an integration test to the GitHub workflows which
starts MySQL server in a container, which allows the backend to be
tested. This ignores several "issues" that exist with the tests, such as
the fact that the test is non-idempotent, and fails to run twice.