Simplify MySQL delete statement
Currently the delete statement in the MySQL backend is a complicated
multiple-table statement, which is a proprietary MySQL extension to the
SQL language. Furthermore, we have observed issues with this delete
statement deadlocking with itself (when two statements run for the same
command row) due to lock-upgrading issues that occur between the query
and delete steps.

I have changed this DELETE statement to be an ordinary delete, with two
simple select statements (with no joins). These select statements may
not have to be annotated `FOR UPDATE`, but I wanted to err on the side
of avoiding deadlocks.

Testing:
* Observe that MySQL integration tests pass
TODO:
* Make sure that deadlocks no longer occur.