Prevent deadlock during command deletion in MySQL
NanoMDM allows for a single command to be queued for multiple devices.
Since #48 landed, both queued commands and the command data itself is
deleted after a device in finished with a command.
However, I have observed deadlocks which occur when multiple devices
trigger deletion for a shared command at once.
This can be avoided by placing a record lock on the command row before
attempting a delete, so that only one device's enrollment queue can be
cleared at a time.

This should have minimal impacts on performance as it will only
serialize deletes to a single command. Furthermore, as the lock on the
command record must be taken anyway, this does not add any additional
locking to the transaction.

Testing:
I have tested this, and have confirmed that it prevents deadlocks in
NanoMDM. Furthermore, I have tested that commands are still delivered
and responded to.