Convert RetrievePushCert to sqlc
I have to cast types because in schema.sql cert_pem and key_pem are TEXT, but when referenced in code are byte[] so sqlc generates them as strings in code. I think changing them to BLOBs in the schema would "fix" this, but not sure about the unintended changes that could make.

Signed-off-by: andrew s <andrew@openedmac.com>